Refactor into packages. split tests. simplify func naming (#27)
* move listener and publisher in separate package
* remove all prefix on option funcs
* move integration test in integration folder
* refactor integration test in 3 suite, 1 for each auth type and reuse the test funcs
* return error from jwt, instead of the token

* remove old code
* make tests share token provider because IMDS barfs on concurrent request :(
* delete unused code
* run tests in parallel where possible
* update readme
