+++ Dockerfile
@@ -1,5 +1,5 @@
-FROM golang
+FROM golang:1.14
 WORKDIR pubsub
 COPY . .
 RUN touch .env
-ENTRYPOINT ["go", "test", "-v"]
+ENTRYPOINT ["go", "test", "-v", "./integration"]
