+++ internal/aad/jwt.go
@@ -222,7 +222,7 @@ func (c *TokenProviderConfiguration) NewServicePrincipalToken() (*adal.ServicePr
 		return nil, fmt.Errorf("failed to get oauth token from MSI: %v", err)
 	}
 	if err := spToken.Refresh(); err != nil {
-		return nil, fmt.Errorf("failed to refersh token: %v", spToken)
+		return nil, fmt.Errorf("failed to refresh token: %w", err)
 	}
 	return spToken, nil
 }
